requirements.txt
asgiref==3.8.1
Django==3.2.12
django-cors-headers==4.4.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.28.2
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
gunicorn==23.0.0
python-dateutil==2.9.0
pyauth @ git+https://github.com/ashokkumarta/pyauth.git@v1.0.2